refactor more top level files
