+++ packages/expo-cli/src/analytics/StatusEventEmitter.ts
