+++ packages/expo-cli/src/utils/update.ts
